Tiled tiles import with correct orientation
